+++ HousingSearchListener.Tests/E2E/GetPersonListRequest.cs
@@ -1,0 +1,24 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace HousingSearchListener.Tests.E2E
+{
+    public class GetPersonListRequest
+    {
+        private const int DefaultPageSize = 12;
+
+        [FromQuery(Name = "searchText")]
+        public string SearchText { get; set; }
+
+        [FromQuery(Name = "pageSize")]
+        public int PageSize { get; set; } = DefaultPageSize;
+
+        [FromQuery(Name = "page")]
+        public int Page { get; set; }
+
+        [FromQuery(Name = "sortBy")]
+        public string SortBy { get; set; }
+
+        [FromQuery(Name = "isDesc")]
+        public bool IsDesc { get; set; }
+    }
+}
